Review : - fix naming - move events in core - remove action.js, containers (useless)
