+++ src/main/core/Event/Security/ForgotPasswordEvent.php
@@ -9,22 +9,19 @@
  * file that was distributed with this source code.
  */
 
-namespace Claroline\LogBundle\Event\Security;
+namespace Claroline\CoreBundle\Event\Security;
 
-use Claroline\CoreBundle\Entity\Role;
 use Claroline\CoreBundle\Entity\User;
 use Symfony\Contracts\EventDispatcher\Event;
 use Symfony\Contracts\Translation\TranslatorInterface;
 
-class RemoveRoleEvent extends Event
+class ForgotPasswordEvent extends Event
 {
     private $user;
-    private $role;
 
-    public function __construct(User $user, Role $role)
+    public function __construct(User $user)
     {
         $this->user = $user;
-        $this->role = $role;
     }
 
     public function getUser(): User
@@ -32,13 +29,8 @@ class RemoveRoleEvent extends Event
         return $this->user;
     }
 
-    public function getRole()
-    {
-        return $this->role;
-    }
-
     public function getMessage(TranslatorInterface $translator)
     {
-        return $translator->trans('removeRole', ['username' => $this->user->getUsername(), 'role' => $this->role->getName()], 'security');
+        return $translator->trans('forgotPassword', ['username' => $this->user->getUsername()], 'security');
     }
 }
