+++ src/main/core/Event/Security/NewPasswordEvent.php
@@ -9,13 +9,13 @@
  * file that was distributed with this source code.
  */
 
-namespace Claroline\LogBundle\Event\Security;
+namespace Claroline\CoreBundle\Event\Security;
 
 use Claroline\CoreBundle\Entity\User;
 use Symfony\Contracts\EventDispatcher\Event;
 use Symfony\Contracts\Translation\TranslatorInterface;
 
-class ValidateEmailEvent extends Event
+class NewPasswordEvent extends Event
 {
     private $user;
 
@@ -29,8 +29,8 @@ class ValidateEmailEvent extends Event
         return $this->user;
     }
 
-    public function getMessage(TranslatorInterface $translator): string
+    public function getMessage(TranslatorInterface $translator)
     {
-        return $translator->trans('validateEmail', ['username' => $this->user->getUsername()], 'security');
+        return $translator->trans('newPassword', ['username' => $this->user->getUsername()], 'security');
     }
 }
