+++ src/main/core/Event/Security/UserEnableEvent.php
@@ -9,21 +9,19 @@
  * file that was distributed with this source code.
  */
 
-namespace Claroline\LogBundle\Event\Security;
+namespace Claroline\CoreBundle\Event\Security;
 
 use Claroline\CoreBundle\Entity\User;
 use Symfony\Contracts\EventDispatcher\Event;
 use Symfony\Contracts\Translation\TranslatorInterface;
 
-class ViewAsEvent extends Event
+class UserEnableEvent extends Event
 {
     private $user;
-    private $role;
 
-    public function __construct(User $user, $role)
+    public function __construct(User $user)
     {
         $this->user = $user;
-        $this->role = $role;
     }
 
     public function getUser(): User
@@ -31,13 +29,8 @@ class ViewAsEvent extends Event
         return $this->user;
     }
 
-    public function getRole()
-    {
-        return $this->role;
-    }
-
     public function getMessage(TranslatorInterface $translator)
     {
-        return $translator->trans('viewAs', ['username' => $this->user->getUsername(), 'role' => $this->role], 'security');
+        return $translator->trans('userEnable', ['username' => $this->user->getUsername()], 'security');
     }
 }
