+++ src/main/log/Resources/config/routing.yml
@@ -1,5 +1,5 @@
 claro_log_new_api:
-    resource: "@ClarolineLogBundle/Controller/APINew/"
+    resource: "@ClarolineLogBundle/Controller/"
     prefix: apiv2
     type: api
     options:
