--- src/main/log/Resources/modules/administration/logs/containers/tool.jsx
@@ -1,13 +1,0 @@
-import {connect} from 'react-redux'
-
-import {withRouter} from '#/main/app/router'
-
-import {LogsTool as LogsToolComponent} from '#/main/log/administration/logs/components/tool'
-
-const LogsTool = withRouter(
-  connect()(LogsToolComponent)
-)
-
-export {
-  LogsTool
-}
