--- src/main/log/Resources/modules/administration/logs/store/actions.js
@@ -1,1 +1,0 @@
-export const actions = {}
