+++ src/main/log/Resources/modules/administration/logs/store/index.js
@@ -1,9 +1,7 @@
-import {actions} from '#/main/log/administration/logs/store/actions'
 import {reducer} from '#/main/log/administration/logs/store/reducer'
 import {selectors} from '#/main/log/administration/logs/store/selectors'
 
 export {
-  actions,
   reducer,
   selectors
 }
