+++ src/main/log/Resources/modules/administration/logs/store/reducer.js
@@ -5,13 +5,13 @@ import {makeListReducer} from '#/main/app/content/list/store/reducer'
 import {selectors} from '#/main/log/administration/logs/store/selectors'
 
 const reducer = combineReducers({
-  securityLogs: makeListReducer(selectors.LIST_NAME, {
+  security: makeListReducer(selectors.LIST_NAME, {
     sortBy: {property: 'date', direction: -1}
   }),
-  messageLogs: makeListReducer(selectors.MESSAGE_NAME, {
+  message: makeListReducer(selectors.MESSAGE_NAME, {
     sortBy: {property: 'date', direction: -1}
   }),
-  functionalLogs: makeListReducer(selectors.FUNCTIONAL_NAME, {
+  functional: makeListReducer(selectors.FUNCTIONAL_NAME, {
     sortBy: {property: 'date', direction: -1}
   })
 })
