+++ src/main/log/Resources/modules/administration/logs/store/selectors.js
@@ -1,7 +1,7 @@
 const STORE_NAME = 'logs'
-const LIST_NAME = STORE_NAME + '.securityLogs'
-const MESSAGE_NAME = STORE_NAME + '.messageLogs'
-const FUNCTIONAL_NAME = STORE_NAME + '.functionalLogs'
+const LIST_NAME = STORE_NAME + '.security'
+const MESSAGE_NAME = STORE_NAME + '.message'
+const FUNCTIONAL_NAME = STORE_NAME + '.functional'
 
 const store = (state) => state[STORE_NAME]
