+++ src/plugin/cursus/Manager/EventManager.php
@@ -14,6 +14,7 @@ namespace Claroline\CursusBundle\Manager;
 use Claroline\AppBundle\Event\StrictDispatcher;
 use Claroline\AppBundle\Persistence\ObjectManager;
 use Claroline\CoreBundle\Entity\User;
+use Claroline\CoreBundle\Event\CatalogEvents\MessageEvents;
 use Claroline\CoreBundle\Event\SendMessageEvent;
 use Claroline\CoreBundle\Manager\PlanningManager;
 use Claroline\CoreBundle\Manager\Template\TemplateManager;
@@ -26,7 +27,6 @@ use Claroline\CursusBundle\Event\Log\LogSessionEventGroupRegistrationEvent;
 use Claroline\CursusBundle\Event\Log\LogSessionEventGroupUnregistrationEvent;
 use Claroline\CursusBundle\Event\Log\LogSessionEventUserRegistrationEvent;
 use Claroline\CursusBundle\Event\Log\LogSessionEventUserUnregistrationEvent;
-use Claroline\LogBundle\Event\CatalogEvents\MessageEvents;
 use Symfony\Component\EventDispatcher\EventDispatcherInterface;
 use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
 use Symfony\Component\Security\Core\Authentication\Token\Storage\TokenStorageInterface;
