+++ src/plugin/tag/Resources/translations/tools.en.json
@@ -1,4 +1,4 @@
 {
     "claroline_tag_admin_tool": "Tags management",
-    "logs": "Logs management"
+    "logs": "Logs"
 }
