+++ src/plugin/tag/Resources/translations/tools.fr.json
@@ -1,4 +1,4 @@
 {
     "claroline_tag_admin_tool": "Gestion des Tags",
-    "logs": "Gestion des Logs"
+    "logs": "Logs"
 }
