+++ src/plugin/tag/Resources/translations/tools.nl.json
@@ -1,4 +1,4 @@
 {
     "claroline_tag_admin_tool": "De tags beheren",
-    "logs": "De logs beheren"
+    "logs": "Logs"
 }
